refactor(biome): update configuration
Add Version Control System (VCS) configuration and adjust assist
configuration. At the moment, biome only supports Git for the VCS. With
this update, biome will ignore all files that are not tracked by Git.
The adjustments to the assist configuration turn the organize imports
functionality on by default, and update the sorting of import
identifiers to lexicographic sorting.